# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sayq POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get today's sales summary
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /dashboard/top-products:
    get:
      operationId: getTopProducts
      tags: [dashboard]
      summary: Get top selling products
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
        - name: branchId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Top products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopProduct"
  /dashboard/sales-chart:
    get:
      operationId: getSalesChart
      tags: [dashboard]
      summary: Get sales data for chart (last 7 days)
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Sales chart data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalesChartPoint"
  /dashboard/cashier-performance:
    get:
      operationId: getCashierPerformance
      tags: [dashboard]
      summary: Get cashier performance stats
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Cashier performance
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CashierPerformance"
...
  /ingredients:
    get:
      operationId: listIngredients
      tags: [ingredients]
      summary: List ingredients (raw materials) with live stock
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of ingredients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ingredient"
    post:
      operationId: createIngredient
      tags: [ingredients]
      summary: Create an ingredient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngredientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ingredient"
  /ingredients/{id}:
    patch:
      operationId: updateIngredient
      tags: [ingredients]
      summary: Update an ingredient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngredientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ingredient"
    delete:
      operationId: deleteIngredient
      tags: [ingredients]
      summary: Delete an ingredient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: Get live inventory for a branch
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
  /inventory/low-stock:
    get:
      operationId: getLowStock
      tags: [inventory]
      summary: Get inventory items below the low-stock threshold
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
        - name: threshold
          in: query
          schema:
            type: number
            default: 200
      responses:
        "200":
          description: Low stock items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
  /stock-adjustments:
    get:
      operationId: listStockAdjustments
      tags: [inventory]
      summary: List stock adjustment history
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Adjustment history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockAdjustment"
    post:
      operationId: createStockAdjustment
      tags: [inventory]
      summary: Record a stock adjustment (in/out/loss); 'in' triggers moving-average COGS
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockAdjustmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockAdjustment"
...
  /dashboard/financial:
    get:
      operationId: getFinancialReport
      tags: [dashboard]
      summary: Get gross revenue, COGS and gross profit
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
        - name: days
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Financial report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialReport"
components:
  schemas:
...
    DashboardSummary:
      type: object
      required: [todayRevenue, todayOrders, totalProducts, lowStockCount, todayRevenueDiff, todayOrdersDiff]
      properties:
        todayRevenue:
          type: number
        todayOrders:
          type: integer
        totalProducts:
          type: integer
        lowStockCount:
          type: integer
        todayRevenueDiff:
          type: number
        todayOrdersDiff:
          type: number
    TopProduct:
      type: object
      required: [productId, productName, totalSold, totalRevenue]
      properties:
        productId:
          type: integer
        productName:
          type: string
        totalSold:
          type: integer
        totalRevenue:
          type: number
    SalesChartPoint:
      type: object
      required: [date, revenue, orders]
      properties:
        date:
          type: string
        revenue:
          type: number
        orders:
          type: integer
    CashierPerformance:
      type: object
      required: [cashierId, cashierName, totalOrders, totalRevenue]
      properties:
        cashierId:
          type: integer
        cashierName:
          type: string
        totalOrders:
          type: integer
        totalRevenue:
          type: number
...
    Ingredient:
      type: object
      required: [id, branchId, name, unit, costPricePerUnit, minimalStock, currentStock]
      properties:
        id:
          type: integer
        branchId:
          type: integer
        name:
          type: string
        unit:
          type: string
        costPricePerUnit:
          type: number
        minimalStock:
          type: number
        currentStock:
          type: number
    IngredientInput:
      type: object
      required: [branchId, name, unit]
      properties:
        branchId:
          type: integer
        name:
          type: string
          minLength: 1
        unit:
          type: string
        minimalStock:
          type: number
          default: 0
    IngredientUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        unit:
          type: string
        minimalStock:
          type: number
...
    InventoryItem:
      type: object
      required: [itemType, itemId, name, unit, currentStock]
      properties:
        itemType:
          type: string
          enum: [ingredient, semi_finished]
        itemId:
          type: integer
        name:
          type: string
        unit:
          type: string
        currentStock:
          type: number
        minimalStock:
          type: ["number", "null"]
        costPricePerUnit:
          type: number
    StockAdjustment:
      type: object
      required: [id, branchId, itemType, itemId, adjustmentType, quantity, createdAt]
      properties:
        id:
          type: integer
        branchId:
          type: integer
        itemType:
          type: string
        itemId:
          type: integer
        itemName:
          type: string
        adjustmentType:
          type: string
          enum: [in, out, loss]
        quantity:
          type: number
        purchasePriceTotal:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    StockAdjustmentInput:
      type: object
      required: [branchId, itemType, itemId, adjustmentType, quantity]
      properties:
        branchId:
          type: integer
        itemType:
          type: string
          enum: [ingredient, semi_finished]
        itemId:
          type: integer
        adjustmentType:
          type: string
          enum: [in, out, loss]
        quantity:
          type: number
          minimum: 0
        purchasePriceTotal:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
...
    FinancialReport:
      type: object
      required: [grossRevenue, totalCogs, grossProfit, grossMarginPct]
      properties:
        grossRevenue:
          type: number
        totalCogs:
          type: number
        grossProfit:
          type: number
        grossMarginPct:
          type: number